+++ tests/fake/test_fake_content_type_ids.py
@@ -6,7 +6,7 @@ def test_default_content_type_ids():
     controller = FakeController()
     client = controller.client
 
-    type_ids = client.get_content_type_ids().result()
+    type_ids = client.get_content_type_ids()
 
     # Type IDs should include this common content types.
     assert "rpm" in type_ids
@@ -24,5 +24,5 @@ def test_set_content_type_ids():
 
     controller.set_content_type_ids(["a", "b", "c"])
 
-    type_ids = client.get_content_type_ids().result()
+    type_ids = client.get_content_type_ids()
     assert sorted(type_ids) == ["a", "b", "c"]