+++ tests/fake/test_fake_delete.py
@@ -9,7 +9,7 @@ def test_can_delete():
     controller.insert_repository(Repository(id="repo2"))
 
     client = controller.client
-    repo1 = client.get_repository("repo1").result()
+    repo1 = client.get_repository("repo1")
 
     # Call to delete should succeed
     delete_f = repo1.delete()
@@ -42,8 +42,8 @@ def test_delete_missing_repo_succeeds():
     # We want to try deleting the same repo more than once.
     # We can't do this through a single reference since it would be detached
     # on delete, but if we get two handles to the same repo, we can.
-    repo_copy1 = client.get_repository("repo").result()
-    repo_copy2 = client.get_repository("repo").result()
+    repo_copy1 = client.get_repository("repo")
+    repo_copy2 = client.get_repository("repo")
 
     # First delete succeeds, with some tasks
     assert repo_copy1.delete().result()