+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- n/a
+
+## [2.1.0] - 2019-09-10
+
+### Added
+- A `search_distributor` API to search distributors on defined `Criteria`
+- `Matcher.less_than()` matcher to find the results with fields less than
+  the given value
+
 ### Fixed
 - Fixed certain exceptions from requests library (such as SSL handshake errors) not being
   propagated correctly.
@@ -14,9 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0] - 2019-09-09
 
 ### Added
-- A `search_distributor` API to search distributors on defined `Criteria`
-- `Matcher.less_than()` matcher to find the results with fields less than
-  the given value
 - ``Page`` objects may now be directly used as iterables
 
 ### Changed